requirements.txt
numpy==1.25.2
opencv_python==4.7.0.72
Pillow==11.2.1
streamlit==1.20.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
